requirements.txt
fastapi == 0.115.0
python3-discogs-client == 2.7.0
requests == 2.31.0
starlette == 0.38.6
streamlit == 1.39.0
uvicorn == 0.31.0
requests == 2.31.0